+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # master
 
+* Rename RoomDirectory#list to RoomDirectory#all
+* Add full support for playlists API
 * Fix Modlist example not sending messages to the room
 
 ## 0.1.4 / 2012-01-08